refactored using interfaces and generics to make code more extensible

+++ app/src/test/java/dev/andrylat/raqimbek/bankingutils/MonthlyMortgagePaymentTestData.java
@@ -4,4 +4,4 @@ public record MonthlyMortgagePaymentTestData(
     double borrowedAmount,
     double annualInterestRate,
     double numberOfYears,
-    double expectedMonthlyPayment) {}
+    long expectedMonthlyPayment) {}
